% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_updated_on_cran.R
\name{check_last_cran_update}
\alias{check_last_cran_update}
\title{Show weeks since last package update on CRAN}
\usage{
check_last_cran_update()
}
\value{
Invisible \code{NULL}.
}
\description{
Green indicates that enough time since last submission has passed and it's ok
to submit an update, yellow means it's ok, but beware it's not too often, and
red means that you should probably not yet submit an update.
}
\note{
This functions is intended for development purposes only, to ensure
that package updates are not submitted to CRAN within too short intervals.
}
\examples{
check_last_cran_update()
}
