% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status_on_cran.R
\name{check_status_on_cran}
\alias{check_status_on_cran}
\title{Show CRAN check status for easystats-packages}
\usage{
check_status_on_cran()
}
\value{
\code{TRUE} if any of the easystats-packages show check-failures on CRAN,
\code{FALSE} otherwise.
}
\description{
Show CRAN check status for easystats-packages
}
\note{
This functions is intended for maintainance-purposes, to timely detect
any check-issues on CRAN for easystats-packages.
}
\examples{
if (interactive()) {
  check_status_on_cran()
}
}
