% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRAN_checks.R
\name{CRAN_checks}
\alias{CRAN_checks}
\title{Show CRAN check status for easystats-packages}
\usage{
CRAN_checks()
}
\value{
\code{TRUE} if any of the easystats-packages show check-failures on CRAN,
\code{FALSE} otherwise.
}
\description{
Show CRAN check status for easystats-packages
}
\note{
This functions is intended for maintainance-purposes, to timely detect
any check-issues on CRAN for easystats-packages.
}
\examples{
if (interactive()) {
  CRAN_checks()
}
}
