% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats-package.R
\docType{package}
\name{easystats-package}
\alias{easystats-package}
\alias{easystats}
\title{easystats: Framework for Easy Statistical Modeling, Visualization, and Reporting}
\description{
A meta-package that installs and loads a set of packages from 'easystats'
ecosystem in a single step. This collection of packages provide a unifying
and consistent framework for statistical modeling, visualization, and
reporting. Additionally, it provides articles targeted at instructors for
teaching 'easystats', and a dashboard targeted at new R users for easily
conducting statistical analysis by accessing summary results, model fit
indices, and visualizations with minimal programming.
}
\details{
\code{easystats}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://easystats.github.io/easystats/}
  \item Report bugs at \url{https://github.com/easystats/easystats/issues}
}

}
\author{
\strong{Maintainer}: Daniel Lüdecke \email{d.luedecke@uke.de} (\href{https://orcid.org/0000-0002-8895-3206}{ORCID})

Authors:
\itemize{
  \item Daniel Lüdecke \email{d.luedecke@uke.de} (\href{https://orcid.org/0000-0002-8895-3206}{ORCID})
  \item Dominique Makowski \email{dom.makowski@gmail.com} (\href{https://orcid.org/0000-0001-5375-9967}{ORCID})
  \item Mattan S. Ben-Shachar \email{matanshm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-4287-4801}{ORCID})
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID})
  \item Brenton M. Wiernik \email{brenton@wiernik.org} (\href{https://orcid.org/0000-0001-9560-6336}{ORCID})
  \item Etienne Bacher \email{etienne.bacher@protonmail.com} (\href{https://orcid.org/0000-0002-9271-5075}{ORCID})
  \item Rémi Thériault \email{remi.theriault@mail.mcgill.ca} (\href{https://orcid.org/0000-0003-4315-6788}{ORCID})
}

}
\keyword{internal}
