% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats_methods.R
\name{easystats_arguments}
\alias{easystats_arguments}
\title{List all arguments used in easystats packages}
\usage{
easystats_arguments(packages = "all", flatten = FALSE)
}
\arguments{
\item{packages}{String, name of the easystats-package, or \code{"all"} to include
all packages.}

\item{flatten}{If \code{TRUE}, returns only one element and each argument only once.}
}
\value{
A named list with argument names used in the different packages.
}
\description{
This function returns a list with all arguments that are used across all
easystats packages
}
