% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats_citations.R
\name{easystats_citations}
\alias{easystats_citations}
\title{Reports citations for easystats publications}
\usage{
easystats_citations()
}
\value{
A data frame of 4 columns: title, journal, year, and cites.
}
\description{
This function reports the total number of Google Scholar citations
for \code{easystats} publications through the \code{scholar} package.
}
\examples{
\dontshow{if (insight::check_if_installed("scholar", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
easystats_citations()
\dontshow{\}) # examplesIf}
}
