% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats_downloads.R
\name{easystats_downloads}
\alias{easystats_downloads}
\title{Reports download statistics for easystats packages}
\usage{
easystats_downloads(from = "2019-02-26", sort_by = "total", length = 30)
}
\arguments{
\item{from}{Starting date for the download statistics. Defaults to
\code{"2019-02-26"}, which is when the first easystats package was released on CRAN.}

\item{sort_by}{Name of the column that should be used for sorting. Can be
\code{"package"}, \code{"total"}, or \code{"monthly"}. Use \code{NULL} for no sorting.
Defaults to \code{"total"} (total downloads).}

\item{length}{Numeric, maximum length of the returned string. If not
\code{NULL}, will shorten the string to a maximum \code{length}, however,
it will not truncate inside words. I.e. if the string length happens to be
inside a word, this word is removed from the returned string, so the
returned string has a \emph{maximum} length of \code{length}, but might be
shorter.}
}
\value{
A data frame with three columns: package, total downloads, and
average monthly downloads.
}
\description{
This function reports download statistics for \code{easystats} packages from CRAN
through the \code{cranlogs} package.
}
\examples{
\dontshow{if (insight::check_if_installed("cranlogs", quietly = TRUE)) withAutoprint(\{ # examplesIf}
\dontrun{
easystats_downloads()
}
\dontshow{\}) # examplesIf}
}
