% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats_model_dashboard.R
\name{easystats_model_dashboard}
\alias{easystats_model_dashboard}
\title{Generate a regression model summary from \code{{easystats}}}
\usage{
easystats_model_dashboard(
  model,
  output_file = "easydashboard.html",
  output_dir = getwd(),
  rmd_dir = system.file("templates/easydashboard.Rmd", package = "easystats")
)
}
\arguments{
\item{model}{A regression model object.}

\item{output_file}{A string specifying the file name in \code{rmarkdown::render()}.
Default is \code{"easydashboard.html"}.}

\item{output_dir}{A string specifying the path to the output directory for
report in \code{rmarkdown::render()}. Default is to use the working directory.}

\item{rmd_dir}{A string specifying the path to the directory containing the
RMarkdown template file. By default, package uses the template shipped with
the package installation (\code{inst/templates/easydashboard.Rmd}).}
}
\value{
An HTML dashboard.
}
\description{
A dashboard containing the following details for the entered regression model:
\itemize{
\item tabular summary of parameter estimates
\item a dot-and-whisker plot for parameter estimates
\item tabular summary of indices for the quality of model fit
\item a collection of models for checking model assumptions
\item a text report
}
}
\examples{
if (FALSE) {
  mod <- lm(wt ~ mpg, mtcars)
  easystats_model_dashboard(mod)
}

}
