% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easystats_packages}
\alias{easystats_packages}
\title{List all packages in the easystats ecosystem}
\usage{
easystats_packages()
}
\value{
A character vector
}
\description{
List all packages in the easystats ecosystem
}
\examples{
easystats_packages()
}
