% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{easystats_update}
\alias{easystats_update}
\title{Update easystats-packages from CRAN, if necessary.}
\usage{
easystats_update()
}
\description{
Update easystats-packages from CRAN, if necessary.
}
