% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats_whatsnew.R
\name{easystats_whatsnew}
\alias{easystats_whatsnew}
\title{List all news from development versions}
\usage{
easystats_whatsnew(packages = "all")
}
\arguments{
\item{packages}{String, name of the easystats-package, or \code{"all"} to include
all packages.}
}
\value{
A table (object of class \emph{gt}), printed to the viewer pane.
}
\description{
This function returns a list with all news from the latest development
packages' versions that are not yet on CRAN.
}
