% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{easystats_zen}
\alias{easystats_zen}
\title{Welcome to the easyverse}
\usage{
easystats_zen()
}
\value{
A reassuring message.
}
\description{
Welcome to the easyverse
}
\examples{
easystats_zen()

}
