% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyStats.R
\name{freqTable}
\alias{freqTable}
\title{freqTable}
\usage{
freqTable(x)
}
\arguments{
\item{x}{vector. numbers to make frequency table.}
}
\value{
result table.
}
\description{
Make Frequency Table
}
\examples{
freqTable(x = x.vec)
}
