% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{install_easystats_dev}
\alias{install_easystats_dev}
\title{Install the dev branches of easystats from github}
\usage{
install_easystats_dev()
}
\description{
Install the dev branches of easystats from github
}
