% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{install_easystats_latest}
\alias{install_easystats_latest}
\title{Install the easystats suite from github}
\usage{
install_easystats_latest()
}
\description{
Install the easystats suite from github
}
