% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{install_latest}
\alias{install_latest}
\title{Install the easystats suite from GitHub or CRAN}
\usage{
install_latest(source = c("github", "cran"))
}
\arguments{
\item{source}{Character. Either `"github"` or `"cran"`. If `"cran"`, packages will be installed from the default CRAN mirror returned by `getOption("repos")$CRAN`.}
}
\description{
This function can be used to install all the easystats packages from either
GitHub (for the latest develolpment versions) or from CRAN (for the current
version on CRAN), If installed from GitHub, packages will be installed from
the stable branch (master/main) for each package.
}
