% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{on_CRAN}
\alias{on_CRAN}
\title{Show weeks since last package update on CRAN}
\usage{
on_CRAN()
}
\description{
Green indicates that enough time since last submission has passed and it's ok to submit an update, yellow means it's ok, but beware it's not too often, and red means that you should probably not yet submit an update.
}
