% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{easystats_update}
\alias{easystats_update}
\title{Update easystats-packages and its dependencies from CRAN, if necessary.}
\usage{
easystats_update(which = c("all", "core", "deps"))
}
\arguments{
\item{which}{String, indicates whether easystats-packages (\code{which = "core"}), dependencies (\code{which = "deps"}) or both (\code{which = "all"}) should be checked for available updates.}
}
\description{
Update easystats-packages and its dependencies from CRAN, if necessary.
}
