% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats_update.R
\name{easystats_update}
\alias{easystats_update}
\title{Update \emph{easystats}-packages and its dependencies from CRAN, if necessary.}
\usage{
easystats_update(which = c("all", "core", "deps"))
}
\arguments{
\item{which}{String, indicates whether easystats-packages (\code{which = "core"}),
dependencies (\code{which = "deps"}) or both (\code{which = "all"}) should be
checked for available updates.}
}
\value{
Invisible \code{NULL}.
}
\description{
Update \emph{easystats}-packages and its dependencies from CRAN, if necessary.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# check which local easystats-packages (and their dependencies)
# are out of date and install updates from CRAN
easystats_update()

# update only easystats core-packages
easystats_update("core")
\dontshow{\}) # examplesIf}
}
