% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{install_easystats_latest}
\alias{install_easystats_latest}
\alias{easystats_install_latest}
\title{Install the easystats suite from github}
\usage{
install_easystats_latest()

easystats_install_latest()
}
\description{
This function can be used to install all the easystats package from github,
either from the master/main branch (the stable one) with
\code{install_easystats_latest()}.
}
