% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_easystats.R
\name{install_latest}
\alias{install_latest}
\title{Install the easystats suite from github}
\usage{
install_latest()
}
\description{
This function can be used to install all the easystats packages from GitHub,
either from the master/main branch (the stable one) with
\code{install_easystats_latest()}.
}
