% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_suggested.R
\name{install_suggested}
\alias{install_suggested}
\title{Download all suggested packages}
\usage{
install_suggested(package = NULL)
}
\arguments{
\item{package}{If \code{NULL}, all suggested packages for all 'easystats'
packages will be installed. If a specific package(s) are specified, only
suggested packages for those package(s) will be installed.}
}
\value{
Useful only for its side-effect of installing the needed packages.
}
\description{
Download all suggested packages
}
\details{
To reduce the dependency load, 'easystats' packages by default will not
download all internally needed packages. It will ask the user to download
them only if they are needed. The current function can help install all
packages a given 'easystats' package might need. For example,
\code{install_suggested("see")}.
}
