% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_suggested.R
\name{install_suggested}
\alias{install_suggested}
\title{Download all suggested packages}
\usage{
install_suggested(package = "easystats")
}
\arguments{
\item{package}{If \code{NULL}, all suggested packages for all 'easystats'
packages will be installed. If specific packages are specified, only
suggested packages for those packages will be installed.}
}
\value{
Useful only for its side-effect of installing the needed packages.
}
\description{
In \code{easystats}, we have a 0-dependency policy, which makes our packages
fairly light and fast to install. However, we rely on many many (many)
packages for testing (at least all the packages for functions that we
support) and some specific features. These "soft dependencies" can be
downloaded at once using this function. This will allow you to fully utilize
100\\% of easystats' functionalities without errors.
}
\details{
To reduce the dependency load, 'easystats' packages by default will not
download all internally needed packages. It will ask the user to download
them only if they are needed. The current function can help install all
packages a given 'easystats' package might need. For example,
\code{install_suggested("see")}.
}
\examples{
\dontrun{
install_suggested("easystats")
}

}
