% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats_model_dashboard.R
\name{model_dashboard}
\alias{model_dashboard}
\title{Generate a regression model summary from \code{{easystats}}}
\usage{
model_dashboard(
  model,
  parameters_args = NULL,
  performance_args = NULL,
  output_file = "easydashboard.html",
  output_dir = getwd(),
  rmd_dir = system.file("templates/easydashboard.Rmd", package = "easystats")
)
}
\arguments{
\item{model}{A regression model object.}

\item{output_file}{A string specifying the file name in \code{rmarkdown::render()}.
Default is \code{"easydashboard.html"}.}

\item{output_dir}{A string specifying the path to the output directory for
report in \code{rmarkdown::render()}. Default is to use the working directory.}

\item{rmd_dir}{A string specifying the path to the directory containing the
RMarkdown template file. By default, package uses the template shipped with
the package installation (\code{inst/templates/easydashboard.Rmd}).}

\item{parameters_arg}{A list of named arguments that are passed down to
\code{parameters::model_parameters()}. For further documentation and details
about the arguments can be found \href{https://easystats.github.io/parameters/reference/model_parameters.html}{here}.
See also 'Examples'.}

\item{performance_arg}{A list of named arguments that are passed down to
\code{performance::model_performance()}. For further documentation and details
about the arguments can be found \href{https://easystats.github.io/performance/reference/model_performance.html}{here}.
See also 'Examples'.}
}
\value{
An HTML dashboard.
}
\description{
A dashboard containing the following details for the entered regression model:
\itemize{
\item tabular summary of parameter estimates
\item a dot-and-whisker plot for parameter estimates
\item tabular summary of indices for the quality of model fit
\item a collection of models for checking model assumptions
\item a text report
}
}
\examples{
if (FALSE) {
  mod <- lm(wt ~ mpg, mtcars)
  model_dashboard(mod)

  # standardize coefficients
  model_dashboard(mod, parameters_args = list(standardize = "refit"))

  # only show selected performance metrics
  model_dashboard(mod, performance_args = list(metrics = c("AIC", "RMSE")))
}

}
