% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easystats_citations.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{print_html}
\alias{print_md}
\alias{display}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{insight}{\code{\link[insight]{display}}, \code{\link[insight:display]{print_html}}, \code{\link[insight:display]{print_md}}}
}}

